Version 2.0
Added Sounds for Wins, both big and small. Added bonus feature where all wins are 10x that of the normal game. Added sound for bonus start and when bonus symbol drops as well as popups for wins, bonus start, bonus wins with animations as well as a goldglow to indicate being in the bonus